Moved out auth logic
