+++ auth/consts/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+	AUTHSECRET : 'shhhhh'
+}
